move getJsonFilePath to FileHelper
